Automatically update csharp client (ref: Cosmo-Tech/cosmotech-api@48986e28d4e357fc8b88ae823f190825830bc944)
===
Adjust Argo Controller settings for cost optimization

This allows to limit the total number of pods and PVCs created
by Argo, after a certain time has elapsed.
Because Workflows are archived, it is safe to delete them
as needed.

Reduce logging clutter when a Workflow is not found by Argo

Log the stack trace only when TRACE'ing.
===

Co-authored-by: rm3l <rm3l@users.noreply.github.com>