Automatically update csharp client (ref: Cosmo-Tech/cosmotech-api@568cc4db7ad681c775e0099e6111afa50633dbdf)
===
Drop useless dependency to Argo Client in 'scenario'

This is useful only in the 'scenariorun' subproject.

Make sure to always copy sub-projects Detekt reports to the root build dir

This allows the 'lint' job to upload SARIF reports
to GitHub Code Scanning, even if the 'detekt' Gradle task does not succeed.
===

Co-authored-by: rm3l <rm3l@users.noreply.github.com>