Automatically update csharp client (ref: Cosmo-Tech/cosmotech-api@5d4b6ac7874b1b535fc0102b23202298ca10cdb0)
===
chore(CI): Drop concurrency constraints in the 'doc' Workflow jobs

Due to [1], any previously pending Job or Workflow in the concurrency
group is canceled, making concurrent runs for different PRs to be skipped.
This kind of defeats the whole purpose of ensuring order across runs.
For now, it should not be an issue since we need to ensure order when
attempting to push, which is done only for pushes to the 'main' branch.
Plus, Git will protect against potential issues at push time.

[1] https://github.com/github/feedback/discussions/5435
===

Co-authored-by: rm3l <rm3l@users.noreply.github.com>